access_control_dashboard/requirements.txt
flask==2.0.1
flask-cors==3.0.10
dash==2.9.0
plotly==5.13.0
pandas==1.5.3
pyodbc==4.0.39
python-dotenv==0.19.2
gunicorn==20.1.0 